bugfixes to aggregations
